♻️ Refactor: 파일 구조 변경 및 컴포저블 추출
