+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/component/CameraPreviewBox.kt
@@ -22,7 +22,7 @@ import coil3.compose.AsyncImage
 import com.twix.designsystem.R
 import com.twix.designsystem.theme.GrayColor
 import com.twix.designsystem.theme.TwixTheme
-import com.twix.task_certification.certification.model.CameraPreview
+import com.twix.task_certification.certification.model.camera.CameraPreview
 import com.twix.task_certification.certification.model.CaptureStatus
 import com.twix.task_certification.certification.model.TorchStatus
 import com.twix.ui.extension.noRippleClickable
