+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/model/camera/Camera.kt
@@ -1,9 +1,8 @@
-package com.twix.task_certification.certification.camera
+package com.twix.task_certification.certification.model.camera
 
 import android.net.Uri
 import androidx.camera.core.CameraSelector
 import androidx.lifecycle.LifecycleOwner
-import com.twix.task_certification.certification.model.CameraPreview
 import com.twix.task_certification.certification.model.TorchStatus
 import kotlinx.coroutines.flow.StateFlow
