+++ feature/task-certification/src/main/java/com/twix/task_certification/detail/component/TaskCertificationCardContent.kt
@@ -10,7 +10,7 @@ import com.twix.designsystem.components.photolog.BackgroundCard
 import com.twix.designsystem.components.photolog.ForegroundCard
 import com.twix.domain.model.enums.BetweenUs
 import com.twix.task_certification.detail.contract.TaskCertificationDetailUiState
-import com.twix.task_certification.detail.swipe.SwipeableCard
+import com.twix.task_certification.detail.component.swipe.SwipeableCard
 
 @Composable
 internal fun TaskCertificationCardContent(
