+++ feature/task-certification/src/main/java/com/twix/task_certification/di/TaskCertificationModule.kt
@@ -3,8 +3,8 @@ package com.twix.task_certification.di
 import com.twix.navigation.NavRoutes
 import com.twix.navigation.base.NavGraphContributor
 import com.twix.task_certification.certification.TaskCertificationViewModel
-import com.twix.task_certification.certification.camera.Camera
-import com.twix.task_certification.certification.camera.CaptureCamera
+import com.twix.task_certification.certification.model.camera.Camera
+import com.twix.task_certification.certification.model.camera.CaptureCamera
 import com.twix.task_certification.detail.TaskCertificationDetailViewModel
 import com.twix.task_certification.editor.TaskCertificationEditorViewModel
 import com.twix.task_certification.navigation.TaskCertificationGraph
